+++ clients/rds.d.ts
@@ -4512,7 +4512,7 @@ declare namespace RDS {
     /**
      *  The maximum number of records to include in the response. If more records exist than the specified value, a pagination token called a marker is included in the response. You can use the marker in a later DescribeExportTasks request to retrieve the remaining results.  Default: 100 Constraints: Minimum 20, maximum 100.
      */
-    MaxRecords?: String;
+    MaxRecords?: MaxRecords;
   }
   export interface DescribeGlobalClustersMessage {
     /**